gce/requirements.txt
Flask==1.0.2
google-cloud-storage==1.19.1
gunicorn==19.6.0
oauth2client==3.0.0
google-cloud-storage==1.28.0
google-cloud-tasks==1.5.0
googleapis-common-protos==1.51.0
pymysql
google-cloud-language==1.3.0